# Backend/docs/swagger.yaml
definitions:
...
  errors.APIError:
    properties:
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
  models.AdminUser:
    properties:
      created_at:
        type: string
      department:
        $ref: '#/definitions/models.Department'
      department_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      last_login:
        type: string
      name:
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      status:
        type: integer
      updated_at:
        type: string
    type: object
  models.Department:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      component_path:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      parent_id:
        type: integer
      path:
        type: string
      permission_items:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      permissions:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.Permission:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menu:
        $ref: '#/definitions/models.Menu'
      menu_id:
        type: integer
      role_ids:
        items:
          type: integer
        type: array
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.AdminUser'
        type: array
    type: object
host: localhost:8082
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: 这是GOCOM项目的API文档，提供了详细的接口说明和使用方法。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GOCOM API
  version: "1.0"
paths:
...
  /api/role:
    get:
      consumes:
      - application/json
      description: 获取所有角色，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回角色列表、总数、页码和每页数量
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取所有角色
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建一个新的角色
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "201":
          description: 返回创建的角色信息
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: 无效的请求参数
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建角色
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 更新特定的角色
      parameters:
      - description: 更新的角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: 返回更新成功消息和更新后的角色信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的请求参数
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新角色
      tags:
      - 角色管理
  /api/role/{id}:
    delete:
      consumes:
      - application/json
      description: 删除特定的角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 无效的ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除角色
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 通过ID获取特定的角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回角色信息
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: 无效的ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 角色未找到
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据ID获取角色
      tags:
      - 角色管理
  /api/role/{id}/assignedMenus:
    get:
      consumes:
      - application/json
      description: 获取特定角色已分配的菜单ID列表
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回已分配的菜单ID列表
          schema:
            additionalProperties:
              items:
                type: integer
              type: array
            type: object
        "400":
          description: 无效的角色ID
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.APIError'
      summary: 获取角色已分配的菜单
      tags:
      - 角色管理
  /api/role/assignMenus:
    post:
      consumes:
      - application/json
      description: 为特定角色分配一组菜单
      parameters:
      - description: 包含role_id和menu_ids的JSON
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 菜单分配成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.APIError'
      summary: 为角色分配菜单
      tags:
      - 角色管理
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 请输入 'Bearer {token}' 来进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"